pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "huggingface>=0.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "ollama>=0.6.0",
    "openai>=2.3.0",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
]
